Start migrating from render callbacks to constructor callbacks.
This is different from the usual way (like in web frameworks like ReactJS), however this allows us to use signals more easily and makes the system less complex, and it's easier to make reactive.
(It is of course inspired by other web frameworks)

For example, it allows for finer control what part of the rendering logic should update on a signal change, while the previous system just ran everything again.